fixing

+++ stack.h
@@ -1,6 +1,6 @@
 typedef struct _stack Stack;
 
 Stack* newStack(int size);
-void stackPush(Stack* stack, float value);
-float stackPop (Stack* stack);
+void stackPush(Stack* stack, int value);
+int stackPop (Stack* stack);
 void stackPrint(Stack* stack);
